Fix compiler warning

+++ libs/libmath/Primes.h
@@ -61,7 +61,7 @@ class Primes {
 
 		// Sanity checks on indices and values
 		bool in_index_range(size_t index) { return 0 <= index && index <= m_primes.size(); }
-		bool in_prime_range(int input) { return 1 <= input && input <= m_primes.back(); }
+		bool in_prime_range(long long int input) { return 1 <= input && input <= m_primes.back(); }
 
 		std::string m_filename;					// Name of file contiaining list of prime numbers
 		std::vector<long long int> m_primes;	// Vector into which is read the list of prime numbers from the file
